start of a class diagram to get some ideas together.
Here's the main thing: the way I fixed #14 is really cheap. I need a better way to do that. Look at the major regex refactoring I did for UserCommands. I need to do the same for Triggers as well.

The thing is, Triggers and UserCommands could boil down to the same thing. There are only 3 main differences:

- what you say if you can't find a UC/Trig to run
- what you say if you don't have permission to run a UC/Trig
- For UCs you strip out any "mentions", Trigs you don't

So with that, is there a way I can effectively boil it down so I copy the least amount of code and get the flow of everything more simplified? That's the big task here.