Everything left that is not xml commented is either a user command/trigger or something in the db model.
I will be removing entity framework in one of the coming versions of the bot (mostly so I can move to mariaDB), so I can ignore that.

The ChatbotActions are just classes that consume the core classes (which are documented). And because nothing should be using a user command or trigger, it is safe to ignore comments on these as well.

So, I'm done with this. Closes #24.