Execute the suggested command if replied to with "yes"
This previously worked fine for commands with no args, and failed
miserably for those with. We're also using reflection here to bypass an
arg type restraint (we have the suggested command text as a String, but
need to pass a Message, so we just mod the message's Content prop
instead).